removed avatar parameter from registering , fixed app merge conflict , added middleware folder
